+++ src/ngrest/plugins/Textarea.php
@@ -96,7 +96,7 @@ class Textarea extends Plugin
      */
     public function onAssign($event)
     {
-        if (!$this->i18n && $this->encoding) {
+        if (!$this->i18n && $this->encoding && $this->isAttributeWriteable($event)) {
             $event->sender->setAttribute($this->name, Html::encode($event->sender->getAttribute($this->name)));
         }
     }